Satellite/amelioration_satellite_advanced.py
import requests
from requests.adapters import HTTPAdapter
from urllib3.util.retry import Retry
from typing import Dict, List, Tuple, Optional
from datetime import datetime, timedelta
import time

class AdvancedSatelliteModule:
    def __init__(self, client_id: str = None, client_secret: str = None):
        self.client_id = client_id
        self.client_secret = client_secret
        self.access_token = None
        self.token_expiry = None
        
        # Configuration avec fallback
        self.endpoints = self._get_available_endpoints()
        
        # Session HTTP avec retry
        self.session = self._create_session()
        
        # Suivi des quotas
        self.quota_usage = {'today': 0, 'month': 0}
        self.quota_limits = {'daily': 1000, 'monthly': 10000}
    
    def _create_session(self):
        """Créer une session HTTP avec retry et timeout"""
        # TODO: Implémenter une session avec retry
        pass
    
    def _get_available_endpoints(self):
        """Endpoints disponibles avec fallback"""
        # TODO: Définir les endpoints primaires et de fallback
        pass
    
    def get_available_dates_with_fallback(self, bbox: Tuple, start_date: str, end_date: str) -> List[Dict]:
        """Récupérer les dates avec système de fallback"""
        # TODO: Implémenter la logique de fallback
        pass
    
    def _fallback_to_public_api(self, bbox: Tuple, start_date: str, end_date: str) -> List[Dict]:
        """Fallback sur des APIs publiques alternatives"""
        # TODO: Implémenter un fallback sur une API publique
        pass