Satellite/routes_amelioree_satellite_blueprint.py
from flask import Blueprint, request, jsonify, current_app, session, g
from flask_caching import Cache
from functools import wraps
import json
from datetime import datetime

satellite_bp = Blueprint('satellite', __name__, url_prefix='/api/satellite')
cache = Cache(config={'CACHE_TYPE': 'simple'})

def cache_response(timeout=300):
    """Décorateur de cache HTTP"""
    # TODO: Implémenter le décorateur de cache
    pass

@satellite_bp.route('/layers', methods=['GET'])
@cache_response(timeout=3600)
def get_available_layers():
    """API optimisée avec compression"""
    # TODO: Implémenter la route avec gestion du cache
    pass

@satellite_bp.route('/best-layer', methods=['POST'])
def get_best_layer():
    """Recommandation automatique de couche"""
    # TODO: Implémenter la recommandation
    pass

@satellite_bp.route('/health', methods=['GET'])
def health_check():
    """Endpoint de santé pour monitoring"""
    # TODO: Implémenter un endpoint de santé
    pass

# TODO: Ajouter les autres routes avec les améliorations