Satellite/ajout_requirements.txt
# requirements-satellite.txt
redis>=4.5.0
requests>=2.28.0
requests-cache>=1.0.1
flask-caching>=2.0.0
flask-compress>=1.13
cryptography>=40.0.0
numpy>=1.24.0
opencv-python-headless>=4.7.0